Remove SettingState
